SvgDrawingPlan -> SvgSection
