+++ Elements.Serialization.SVG/src/SaveOptions/PdfSaveOptions.cs
@@ -1,0 +1,9 @@
+namespace Elements.Serialization.SVG
+{
+    public class PdfSaveOptions
+    {
+        public float Margin { get; set; } = 15;
+        public float PageWidth { get; set; } = 840;
+        public float PageHeight { get; set; } = 1188;
+    }
+}
